refactor: move simplecrypto to common
- delete unused datastructures package
- add common/ testing in workflows
