+++ apps/client/internal/domain/entities/user.go
@@ -1,7 +1,7 @@
 package entities
 
 import (
-	"eaglechat/apps/client/internal/utils/simplecrypto/rsa"
+	"eaglechat/common/simplecrypto/rsa"
 	"time"
 )
