+++ apps/client/internal/domain/services/middleware.go
@@ -2,7 +2,7 @@ package services
 
 import (
 	"eaglechat/apps/client/internal/domain/entities"
-	"eaglechat/apps/client/internal/utils/simplecrypto/rsa"
+	"eaglechat/common/simplecrypto/rsa"
 )
 
 type Middleware interface {
