+++ apps/client/internal/infrastructure/repositories/sqlite/chats.go
@@ -1,11 +1,10 @@
 package sqliterepository
 
 import (
-	"time"
-
 	"eaglechat/apps/client/internal/domain/entities"
 	"eaglechat/apps/client/internal/domain/repositories"
-	"eaglechat/apps/client/internal/utils/simplecrypto/rsa"
+	"eaglechat/common/simplecrypto/rsa"
+	"time"
 )
 
 // GetAllChatOverviews returns the corresponding ChatOverview object for each chat.
