+++ apps/client/internal/infrastructure/repositories/sqlite/message.go
@@ -2,7 +2,7 @@ package sqliterepository
 
 import (
 	"eaglechat/apps/client/internal/domain/entities"
-	"eaglechat/apps/client/internal/utils/simplecrypto/rsa"
+	"eaglechat/common/simplecrypto/rsa"
 	"time"
 )
