+++ apps/client/internal/infrastructure/repositories/sqlite/profile.go
@@ -4,7 +4,7 @@ import (
 	"database/sql"
 	"eaglechat/apps/client/internal/domain/entities"
 	"eaglechat/apps/client/internal/domain/repositories"
-	"eaglechat/apps/client/internal/utils/simplecrypto/rsa"
+	"eaglechat/common/simplecrypto/rsa"
 )
 
 // SaveOwnProfile saves the full user profile, including the private key, to local storage.
