+++ apps/client/internal/middleware/domain/entities/id_manager_conn.go
@@ -2,7 +2,7 @@ package entities
 
 import (
 	"eaglechat/apps/client/internal/domain/entities"
-	"eaglechat/apps/client/internal/utils/simplecrypto/rsa"
+	"eaglechat/common/simplecrypto/rsa"
 )
 
 type IDManagerConnection interface {
