+++ apps/client/internal/middleware/domain/entities/id_manager_data.go
@@ -1,7 +1,7 @@
 package entities
 
 import (
-	"eaglechat/apps/client/internal/utils/simplecrypto/rsa"
+	"eaglechat/common/simplecrypto/rsa"
 	"net"
 )
