+++ apps/client/internal/middleware/domain/services/id_manager_pool.go
@@ -3,7 +3,7 @@ package services
 import (
 	"eaglechat/apps/client/internal/domain/entities"
 	middleware_entities "eaglechat/apps/client/internal/middleware/domain/entities"
-	"eaglechat/apps/client/internal/utils/simplecrypto/rsa"
+	"eaglechat/common/simplecrypto/rsa"
 )
 
 // IDManagerPool manages listening to new id managers broadcasts in the network
