+++ apps/client/internal/middleware/infrastructure/idmanagerconn/id_manager_conn.go
@@ -2,7 +2,7 @@ package idmanagerconn
 
 import (
 	"eaglechat/apps/client/internal/domain/entities"
-	"eaglechat/apps/client/internal/utils/simplecrypto/rsa"
+	"eaglechat/common/simplecrypto/rsa"
 	"encoding/json"
 	"fmt"
 	"net/http"
