+++ apps/client/internal/middleware/infrastructure/idmanagerconn/query_users.go
@@ -3,7 +3,7 @@ package idmanagerconn
 import (
 	"bytes"
 	"eaglechat/apps/client/internal/domain/entities"
-	"eaglechat/apps/client/internal/utils/simplecrypto/rsa"
+	"eaglechat/common/simplecrypto/rsa"
 	"encoding/json"
 	"fmt"
 	"log"
