+++ apps/client/internal/middleware/infrastructure/idmanagerregisterer/registerer.go
@@ -4,8 +4,8 @@ import (
 	"context"
 	"eaglechat/apps/client/internal/domain/entities"
 	"eaglechat/apps/client/internal/domain/services"
-	"eaglechat/apps/client/internal/utils/simplecrypto/rsa"
 	"eaglechat/common/multicast/implementation"
+	"eaglechat/common/simplecrypto/rsa"
 	"fmt"
 	"log"
 	"time"
