+++ apps/client/internal/ui/controller/registration.go
@@ -3,7 +3,7 @@ package controller
 import (
 	"eaglechat/apps/client/internal/domain/entities"
 	"eaglechat/apps/client/internal/ui/models"
-	"eaglechat/apps/client/internal/utils/simplecrypto/rsa"
+	"eaglechat/common/simplecrypto/rsa"
 	"fmt"
 	"log"
 )
