+++ common/simplecrypto/rsa/serialization_test.go
@@ -1,11 +1,11 @@
 package rsa_test
 
 import (
+	"eaglechat/common/simplecrypto/rsa"
 	"os"
 	"path/filepath"
 	"testing"
 
-	"eaglechat/apps/client/internal/utils/simplecrypto/rsa"
 	"github.com/stretchr/testify/assert"
 )
